Change the way a tree of trees is constructed.
Old way:  Pass in a list of entity set handles, where each set contains
          at least one triangle.  Where single function would:
          1) check for a tag on each set pointing to the OBB tree for that set
          2) build OBB trees for each set that didn't already have one
          3) store OBB tree for each set in tag on set
          4) put each set in the list of contents for it's OBB tree root
          5) build a single tree from the OBB trees of each set

New way:  1) build a tree for each set with separate calls to tool
          2) call new join_trees() method to combine all trees into one.

The new way is more general: it doesn't create or depend on any method
of linking an entity set (e.g. a surface) with an OBB tree or subtree.
The caller can set up these links in whatever way seems most appropriate
for the application.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@904 6162379e-bd28-0410-9a7d-b7f4dcfcad3c