Add functions to read columns of the set description table separately. This allows reading data such as the flags or the offsets in the set content table separately from other per-set values.
Also, when reading really old .h5m files that don't have set parent
links, set the parent offsets for all sets to -1.  This shouldn't
matter because the file will not contain a set parent table anyway,
but -1 is the correct value to indicate that all the sets have no
parents.

 include/mhdf.h |   90 +++++++++++++++++++++++++++
 src/sets.c     |  186 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++-
 2 files changed, 273 insertions(+), 3 deletions(-)


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@990 6162379e-bd28-0410-9a7d-b7f4dcfcad3c