Fix a bug in the SMF reader, providing better resistance to malformed input.
This fixes an unusual condition in which the SMF reader would get stuck
in an infinite loop (if the failbit of ifstream 'smfFile' became true before
the end of file was reached).  In sufficiently unlucky circumstances, this
could cause the MOAB test suite to hang.

It's not obvious to me that MB_FILE_WRITE_ERROR is the correct error to
return on malformed input, but I have used this error code for
consistency with ReadSmf::parse_line().

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4127 6162379e-bd28-0410-9a7d-b7f4dcfcad3c