Add tests verifying that portability of binary data is being handled correctly for .h5m files.
Previously, we didn't have to worry about this because data conversion
was handled implicitly by H5Dread calls.  However, future changes to
the .h5m reader will do data conversion with explicit calls to H5Tconvert
to avoid some parallel performance pitfalls of the HDF5 library.  These
tests verify that the conversion is being done everywhere it should be
and that buffer sizes are being handled correctly when the conversion
involves a change in the size of the type.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4141 6162379e-bd28-0410-9a7d-b7f4dcfcad3c