Fix invalid sharing data for interface elements from ParallelComm::resolve_shared_ents
Code initially guesses at sharing for interface elements as the set of all
processes that share all of the vertices defining the element.  Code existed
to check this after initial handle exchange to verify that it was correct.
Upon identifying extra processes in the sharing list for an element the code 
updated the sharing information for the element.  However, additional state
is not updated correctly for such elements that are still shared by more than
one process.  Update ownership state and interface sets (including possibly
creating new interface sets and removing now empty ones) also.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4545 6162379e-bd28-0410-9a7d-b7f4dcfcad3c