Fix bug writing entity sets to HDF5 file.
The MOAB API is a little bit broken in that MESHSET_ORDERED and MESHSET_SET
are multually excusive and one or the other is always implied.  That is,
MOAB has only two internal set types : "set" and "list".  But the two different
flags allow for four different combinations.  What happens internally
is that the choice of set type is made solely on whether or not
MESHSET_ORDERED is specified (specified implies "list", unspecified implies
"set".)  The MESHSET_SET flag is ignored.  Make HDF5 writer work the same
way (check for MESHSET_ORDERED and ingore MESHSET_SET.)



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4548 6162379e-bd28-0410-9a7d-b7f4dcfcad3c