replace an assert in cubit file reader
asserts are triggered in debug mode only
 This assert should not be ignored in optimized mode,
because if we ignore it,  we are setting some tags on entity handle 0,
which is not what we want.
Instead, we will get out early from that routine;
Entity Handle is 0 in this case because we do not have a geometric mesh set
defined for a geometric entity that is not meshed.