fbigeom exposure to the world
fbigeom needs to be exposed better for mesquite.
Basically, before this, iRel-Def.inc, did not have a proper reference to FBiGeom.
introduce an extra variable, that will be substituted properly.
I called it FBIGEOM_EXTRA_LIB, but maybe there is a better way.
With this fix, mesquite can be configured even if lasso was configured 
with fbigeom. 
So no need anymore to do the following build trick, to make mesquite happy:

This is what I had to do before:
1) configure lasso without fbigeom
2) build, install lasso 
3) configure mesquite with imesh, igeom, irel; build, install
4) go back to lasso, configure with fbigeom, install
5) go back to mesquite, DO NOT reconfigure, but build with the new lasso.
make, install, and everybody was happy

Now, you just do the normal thing: 
1) build lasso with fbigeom, install
2) configure mesquite and be happy