inform the user about changing the writer
this commit was triggered by trying to convert to an exodus file
The writer failed in some particular case, but then moab switched to a
default writer (usually hdf5, for a typical configuration) that worked fine.
The file is saved though with the original extension, and mbsize fails to
read it back.