more robust intersection needed
first thing for intx on sphere, verify that the quads are not degenerate and
that the elements are oriented positively (the area should be positive to
start with)

more work needed on the robustness, not all elements are decomposed
correctly , really big disappointment.
Did I work on the wrong problem or did I think that this is robust enough?